test: 추첨 이벤트 테스트 코드 작성 (#108)
